Update ExampleSourceCode.txt
Source code had an operation that I forgot was not supported by the computer's ALU. The ZERO operation is actually ONE, and A can only be read from outside of the ALU, so I swapped A and B too. (Though, now that I think about it, I might be able to change this if I do something like: if MFB, do ALU PASB, instead of reading from B. Not sure how well this would work though. Nevermind, just checked, those wires are way too dense to do that.)